requirements.txt
python-can
paho-mqtt
numpy
